requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
laces==0.1.1
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.2.0
pillow_heif==0.15.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
urllib3==2.2.1
wagtail==6.0.1
webencodings==0.5.1
Willow==1.8.0
